// src/app/models/costo_representada_por_cebe.model.ts
export interface CostoRepresentadaPorCeBe {
  claseFactura: string;
  moneda: string;
  orgVentas: string;
  fechaCtble: string;
  clienteId: string;
  cbteNro: string;
  cbtePos: string;
  cantidad: number;
  fechaCtble2: string;
  materialId: string;
  materialDesc: string;
  paisId: string;
  ceBe: string;
  datosContrato: string;
  netoSinIva: number;
}